The main question is How unfortunate is for being black in US? We are  make the line graph for black alaves regording to the differen region. The graph shows that there is an increasing trend  for each region in genel. for the region of USA total, there is a dramatical increase in 1927, and around in 1860,it gets the peak.However, after the peak, the number of the black slaves was reduced. Moreover, the line of south region indicates that there was a increasing trend and the variation was proportionately to the level of the series. Finall, it decreased quickly around in 1860.
For the graph of the black free in different regions, the number of the free black people increased dramatically around year 1860, especially for the regions of south and USA toal.

we try to look at the percentage bewteen the free black people and black slaves.
According to the census data,  69% of the black were slaves,and just 31% were free. The black slaves is more than twice if the black free people.In other words, there was a higher chance to be a slave if people are black.
However, we focus on the difference between the black and white people. it is easy to find that hite people were dominant with 85%, only 16 % were black, which including 11% of slaves.

all in all, being an black people is unfortuatly in that situation.